+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [fix] fix navigation for many menu items
 - [fix] update language files
 - [fix] fix button styles
+- [fix] add protected legend in dca
 
 ## [2.18.0](https://github.com/contao-themes-net/mate-theme-bundle/tree/2.18.0) – 2022-05-12